+++ Runtime/Scripts/System/Socket/WebSocket/Core/OpenEventArgs.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace MyVerseXRSDK
 {
-    public class OpenEventArgs : EventArgs
+    internal class OpenEventArgs : EventArgs
     {
         public OpenEventArgs()
         {